requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
bitarray==2.7.3
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.12.1
Django==4.1.7
djangorestframework==3.14.0
drf-yasg==1.21.5
eth-abi==4.0.0b3
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-dataclass==8.5.12
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
parsimonious==0.9.0
protobuf==4.22.1
psycopg2==2.9.5
pycryptodome==3.17
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.2
regex==2022.10.31
requests==2.28.2
rlp==3.0.0
ruamel.yaml==0.17.21
sqlparse==0.4.3
toolz==0.12.0
typing-inspect==0.8.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
web3==6.0.0
websockets==10.4
yarl==1.8.2
